requirements.txt
streamlit==1.32.2
scikit-learn==1.4.2
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.2
jieba==0.42.1
